requirements.txt
streamlit
torch
torchvision
albumentations
pillow
numpy
